created new PassportPopup file for AuthPopup
