+++ apps/consumer-client/pages/examples/group-proof.tsx
@@ -1,6 +1,6 @@
 import {
   constructPassportPcdGetRequestUrl,
-  usePassportResponse,
+  usePassportPopupMessages,
   usePCDMultiplexer,
   usePendingPCD,
   useSemaphorePassportProof,
@@ -23,8 +23,8 @@ import { requestProofFromPassport } from "../../src/util";
  * request a Semaphore Group Membership PCD as a third party developer.
  */
 export default function Page() {
-  const [passportPCDStr, passportPendingPCDStr] = usePassportResponse();
-  const [serverProving, setServerProving] = useState(false);
+  // Populate PCD from either client-side or server-side proving using passport popup
+  const [passportPCDStr, passportPendingPCDStr] = usePassportPopupMessages();
   const [pendingPCDStatus, pendingPCDError, serverPCDStr] = usePendingPCD(
     passportPendingPCDStr,
     PASSPORT_SERVER_URL
@@ -36,6 +36,7 @@ export default function Page() {
   );
 
   const [debugChecked, setDebugChecked] = useState(false);
+  const [serverProving, setServerProving] = useState(false);
 
   return (
     <>
