+++ apps/consumer-client/pages/zuzalu-examples/group-proof.tsx
@@ -1,6 +1,6 @@
 import {
   requestZuzaluMembershipUrl,
-  usePassportResponse,
+  usePassportPopupMessages,
   usePCDMultiplexer,
   usePendingPCD,
   useSemaphorePassportProof,
@@ -21,8 +21,8 @@ import { requestProofFromPassport } from "../../src/util";
  * request a Semaphore Group Membership PCD as a third party developer.
  */
 export default function Page() {
-  const [passportPCDStr, passportPendingPCDStr] = usePassportResponse();
-  const [serverProving, setServerProving] = useState(false);
+  // Populate PCD from either client-side or server-side proving using passport popup
+  const [passportPCDStr, passportPendingPCDStr] = usePassportPopupMessages();
   const [pendingPCDStatus, pendingPCDError, serverPCDStr] = usePendingPCD(
     passportPendingPCDStr,
     PASSPORT_SERVER_URL
@@ -32,6 +32,8 @@ export default function Page() {
     SEMAPHORE_GROUP_URL,
     pcdStr
   );
+
+  const [serverProving, setServerProving] = useState(false);
 
   return (
     <>
