+++ apps/consumer-client/pages/zuzalu-examples/uuid-proof.tsx
@@ -1,7 +1,7 @@
 import {
   requestSignedZuzaluUUIDUrl,
   useFetchParticipant,
-  usePassportResponse,
+  usePassportPopupMessages,
   usePCDMultiplexer,
   usePendingPCD,
   useSemaphoreSignatureProof,
@@ -19,8 +19,8 @@ import { requestProofFromPassport } from "../../src/util";
  * party developer.
  */
 export default function Page() {
-  const [passportPCDStr, passportPendingPCDStr] = usePassportResponse();
-  const [serverProving, setServerProving] = useState(false);
+  // Populate PCD from either client-side or server-side proving using passport popup
+  const [passportPCDStr, passportPendingPCDStr] = usePassportPopupMessages();
   const [pendingPCDStatus, pendingPCDError, serverPCDStr] = usePendingPCD(
     passportPendingPCDStr,
     PASSPORT_SERVER_URL
@@ -40,6 +40,8 @@ export default function Page() {
 
   // Finally, once we have the UUID, fetch the participant data from Passport.
   const { participant } = useFetchParticipant(PASSPORT_SERVER_URL, uuid);
+
+  const [serverProving, setServerProving] = useState(false);
 
   return (
     <>
