+++ packages/passport-interface/src/PCDIntegration.ts
@@ -25,33 +25,6 @@ export function useProof<T extends PCDPackage>(
 }
 
 /**
- * React hook that listens for PCDs and PendingPCDs from a passport popup window
- * using message passing and event listeners.
- */
-export function usePassportResponse() {
-  const [pcdStr, setPCDStr] = useState("");
-  const [pendingPCDStr, setPendingPCDStr] = useState("");
-
-  // Listen for PCDs coming back from the Passport popup
-  useEffect(() => {
-    function receiveMessage(ev: MessageEvent<any>) {
-      // Extensions including Metamask apparently send messages to every page. Ignore those.
-      if (ev.data.encodedPCD) {
-        console.log("Received PCD", ev.data.encodedPCD);
-        setPCDStr(ev.data.encodedPCD);
-      } else if (ev.data.encodedPendingPCD) {
-        console.log(ev.data);
-        setPendingPCDStr(ev.data.encodedPendingPCD);
-      }
-    }
-    window.addEventListener("message", receiveMessage, false);
-    return () => window.removeEventListener("message", receiveMessage);
-  }, []);
-
-  return [pcdStr, pendingPCDStr];
-}
-
-/**
  * React hook that pings server on status of a PendingPCD. Returns a serialized
  * PCD when a completed PCD is returned, or the current status.
  */
